+++ src/cryptocurrency/dto/daily.dto.ts
@@ -1,0 +1,7 @@
+import { DataType } from '../..';
+
+export type DailyDTO = {
+  symbol: string;
+  market: string;
+  datatype?: DataType;
+};