+++ src/cryptocurrency/dto/intraday-response.dto.ts
@@ -1,0 +1,24 @@
+export type IntradayMetadata = {
+  information: string;
+  digitalCurrencyCode: string;
+  digitalCurrencyName: string;
+  marketCode: string;
+  marketName: string;
+  lastRefreshed: string;
+  interval: string;
+  outputSize: string;
+  timeZone: string;
+};
+
+export type IntradayTimeSeries = {
+  open: string;
+  high: string;
+  low: string;
+  close: string;
+  volume: number;
+};
+
+export type IntradayResponseDTO = {
+  metadata: IntradayMetadata;
+  timeSeries: { [key: string]: IntradayTimeSeries };
+};